+++ kernel/src/main/java/org/kframework/kore/convertors/KILtoKORE.java
@@ -33,7 +33,7 @@ import org.kframework.kil.UserList;
 import org.kframework.kore.Attributes;
 import org.kframework.kore.outer.*;
 
-import static org.kframework.kore.Collections.*;
+import static org.kframework.Collections.*;
 import scala.Enumeration.Value;
 import scala.collection.Seq;
