+++ kore/src/main/java/org/kframework/kore/Constructors.java
@@ -2,8 +2,7 @@
 
 package org.kframework.kore;
 
-import static org.kframework.kore.outer.Constructors.*;
-import static org.kframework.kore.Collections.*;
+import static org.kframework.Collections.*;
 
 import java.util.List;
