+++ kore/src/main/scala/org/kframework/collections.scala
@@ -64,3 +64,19 @@ case class ImmutableSet[T](s: collection.Set[T]) extends Collection[T] {
   def foreach(f: T => Unit) = s foreach f
   def iterable: Iterable[T] = s
 }
+
+import org.kframework._
+import collection.JavaConverters._
+import java.util.stream.StreamSupport
+
+object Collections {
+  def immutable[T](s: java.lang.Iterable[T]): Iterable[T] = s.asScala
+  def immutable[T](s: java.util.Set[T]): Set[T] = s.asScala.toSet
+  def immutable[T](s: java.util.List[T]): Seq[T] = s.asScala
+  def immutable[T](s: Array[T]): Seq[T] = s
+
+  def iterable[T](c: Iterable[T]): java.lang.Iterable[T] = c.asJava
+  def stream[T](c: Iterable[T]): java.util.stream.Stream[T] = StreamSupport.stream(c.asJava.spliterator(), false);
+  def stream[T](c: Collection[T]): java.util.stream.Stream[T] = c.stream
+  def iterable[T](c: Collection[T]): java.lang.Iterable[T] = c.iterable.asJava
+}
